move define file to domain dir
