+++ community/domain/validation.py
@@ -1,6 +1,6 @@
 from django.core.exceptions import ValidationError
 
-from community.domain.categories import *
+from community.domain.define import TREE_STRUCTURE
 
 
 class CategoryValidator:
