+++ community/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 from account.models import User
-from community.domain.categories import CategoryDepth1, CategoryDepth2, CategoryDepth3, CategoryDepth4
-from community.domain.validation import CategoryValidator, LikeValidator, PostCategories
+from community.domain.validation import CategoryValidator, LikeValidator
+from community.domain.define import PostCategories, CategoryDepth1, CategoryDepth2, CategoryDepth3, CategoryDepth4
 
 
 class Post(models.Model):
